Ignore unrecognized or additional attributes when parsing Twine archive
This should make the parser a little less likely to break when Twine or a story format updates and adds a new attribute that Spiner doesn't recognize yet.

New errors could be added later, but this is the only one for now of which I am sure Spiner can handle.